backend/services/graphql-service/requirements.txt
# Framework GraphQL
strawberry-graphql[fastapi]==0.219.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Base de datos
psycopg2-binary==2.9.9      # Para PostgreSQL


# Utilidades
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0               # Para llamadas HTTP a otros servicios

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3